refactor `sidecar/connections.ts` into subdir
