+++ src/sidecar/connections/local.ts
@@ -1,0 +1,107 @@
+import {
+  tryToCreateConnection,
+  tryToDeleteConnection,
+  tryToGetConnection,
+  tryToUpdateConnection,
+} from ".";
+import { ContainerListRequest, ContainerSummary, Port } from "../../clients/docker";
+import { Connection, ConnectionSpec } from "../../clients/sidecar";
+import { LOCAL_CONNECTION_ID, LOCAL_CONNECTION_SPEC } from "../../constants";
+import {
+  getLocalSchemaRegistryImageName,
+  getLocalSchemaRegistryImageTag,
+  isDockerAvailable,
+} from "../../docker/configs";
+import { MANAGED_CONTAINER_LABEL } from "../../docker/constants";
+import { getContainersForImage } from "../../docker/containers";
+import { Logger } from "../../logging";
+
+const logger = new Logger("sidecar.connections.local");
+
+/** Create the local {@link Connection} and return it. */
+export async function createLocalConnection(): Promise<Connection> {
+  return await tryToCreateConnection(LOCAL_CONNECTION_SPEC);
+}
+
+/** Get the local {@link Connection} (if it exists). */
+export async function getLocalConnection(): Promise<Connection | null> {
+  return await tryToGetConnection(LOCAL_CONNECTION_ID);
+}
+
+/**
+ * Discover any supported locally-running resources and update the local connection with the
+ * relevant resource configs (for now, just Schema Registry containers, but in the future will
+ * include Kafka and other resources).
+ */
+export async function updateLocalConnection(schemaRegistryUri?: string): Promise<void> {
+  let spec: ConnectionSpec = LOCAL_CONNECTION_SPEC;
+
+  // TODO(shoup): look up Kafka containers here once direct connections are used
+
+  schemaRegistryUri = schemaRegistryUri ?? (await discoverSchemaRegistry());
+  if (schemaRegistryUri) {
+    spec = {
+      ...LOCAL_CONNECTION_SPEC,
+      local_config: {
+        ...LOCAL_CONNECTION_SPEC.local_config,
+        schema_registry_uri: schemaRegistryUri,
+      },
+    };
+  }
+
+  const currentLocalConnection: Connection | null = await getLocalConnection();
+  // inform sidecar if the spec has changed (whether this is a new spec or if we're changing configs)
+  if (!currentLocalConnection) {
+    await tryToCreateConnection(spec);
+  } else if (JSON.stringify(spec) !== JSON.stringify(currentLocalConnection?.spec)) {
+    await tryToUpdateConnection(spec);
+  }
+}
+
+/** Delete the existing local {@link Connection} (if it exists). */
+export async function deleteLocalConnection(): Promise<void> {
+  await tryToDeleteConnection(LOCAL_CONNECTION_ID);
+}
+
+/** Discover any running Schema Registry containers and return the URI to include the REST proxy port. */
+async function discoverSchemaRegistry(): Promise<string | undefined> {
+  const dockerAvailable = await isDockerAvailable();
+  if (!dockerAvailable) {
+    return;
+  }
+
+  const imageRepo = getLocalSchemaRegistryImageName();
+  const imageTag = getLocalSchemaRegistryImageTag();
+  const repoTag = `${imageRepo}:${imageTag}`;
+  const containerListRequest: ContainerListRequest = {
+    all: true,
+    filters: JSON.stringify({
+      ancestor: [repoTag],
+      label: [MANAGED_CONTAINER_LABEL],
+      status: ["running"],
+    }),
+  };
+  const containers: ContainerSummary[] = await getContainersForImage(containerListRequest);
+  if (containers.length === 0) {
+    return;
+  }
+  // we only care about the first container
+  const container: ContainerSummary = containers.filter((c) => !!c.Ports)[0];
+  const ports: Port[] = container.Ports?.filter((p) => !!p.PublicPort) || [];
+  if (!ports || ports.length === 0) {
+    logger.debug("No ports found on Schema Registry container", { container });
+    return;
+  }
+  const schemaRegistryPort: Port | undefined = ports.find((p) => !!p.PublicPort);
+  if (!schemaRegistryPort) {
+    logger.debug("No PublicPort found on Schema Registry container", { container });
+    return;
+  }
+  const restProxyPort = schemaRegistryPort.PublicPort;
+  if (!restProxyPort) {
+    logger.debug("No REST proxy port found on Schema Registry container", { container });
+    return;
+  }
+  logger.debug("Discovered Schema Registry REST proxy port", { schemaRegistryPort });
+  return `http://localhost:${restProxyPort}`;
+}
