+++ src/sidecar/connections/statusUtils.ts
@@ -1,13 +1,13 @@
 // helpers for connection status testing, factored out for test spying
 
-import { reactToCCloudAuthState } from "../authn/ccloudStateHandling";
-import { Connection, ConnectionType } from "../clients/sidecar/models";
-import { connectionStable, environmentChanged } from "../emitters";
-import { Logger } from "../logging";
-import { ConnectionId, EnvironmentId } from "../models/resource";
-import { ConnectionEventBody } from "../ws/messageTypes";
+import { reactToCCloudAuthState } from "../../authn/ccloudStateHandling";
+import { Connection, ConnectionType } from "../../clients/sidecar/models";
+import { connectionStable, environmentChanged } from "../../emitters";
+import { Logger } from "../../logging";
+import { ConnectionId, EnvironmentId } from "../../models/resource";
+import { ConnectionEventBody } from "../../ws/messageTypes";
 
-const logger = new Logger("connectionStatusUtils");
+const logger = new Logger("sidecar.connections.status");
 
 /**
  * Perform side-effects in the UI based on the the connection action (CREATED, DELETED, CONNECTED, etc.)
